Implement blueprint for cabs
