+++ website/templates/home_page.html
@@ -33,7 +33,7 @@
               </li>
               <li>
                 <!-- <i class="ph-fill ph-train"></i> -->
-                <a href="{{ url_for('list_cab') }}">Locomotivas</a>
+                <a href="{{ url_for('cab_blueprint.list_cab') }}">Locomotivas</a>
               </li>
               <li>
                 <!-- <i class="ph ph-arrows-left-right"></i> -->
