+++ website/templates/list_cab.html
@@ -17,13 +17,13 @@
       <article>
         <p>{{cab}}</p>
         <div class="tools">
-          <form id="form" method="POST" action="{{ url_for('edit_cab') }}">
+          <form id="form" method="POST" action="{{ url_for('cab_blueprint.edit_cab') }}">
           <button class="edit-icon" id=""><!--  colocar id-->
             <i class="ph ph-pencil-line"></i>
           </button>
           <input type="hidden" name="edit_id" id="edit_id" value="{{id}}">
           </form>
-          <form id="form" method="POST" action="{{ url_for('try_remove_cab') }}">
+          <form id="form" method="POST" action="{{ url_for('cab_blueprint.try_remove_cab') }}">
           <button class="trash-icon" id=""><!--  colocar id-->
             <i class="ph ph-trash"></i>
             <input type="hidden" name="cab_id" id="cab_id" value="{{id}}">
@@ -36,7 +36,7 @@
     </div>
   </div>
   <div class="credencial">
-    <a href="{{url_for('register_cab')}}">
+    <a href="{{url_for('cab_blueprint.register_cab')}}">
       <button type="submit" >
         Cadastrar locomotivas
       </button>
